feat: ImGui input handling; enum => enum class

+++ Doodle/src/Event/KeyEvent.h
@@ -38,7 +38,7 @@ public:
     std::string ToString() const override
     {
         std::stringstream ss;
-        ss << "KeyPressedEvent: " << m_keyCode << " (" << m_repeatCount << " repeats)";
+        ss << "KeyPressedEvent: " << static_cast<int>(m_keyCode) << " (" << m_repeatCount << " repeats)";
         return ss.str();
     }
 
@@ -57,7 +57,7 @@ public:
     std::string ToString() const override
     {
         std::stringstream ss;
-        ss << "KeyReleasedEvent: " << m_keyCode;
+        ss << "KeyReleasedEvent: " << static_cast<int>(m_keyCode);
         return ss.str();
     }
